Refactor error handling & data fetching & storage

+++ frontend/src/pages/singleJob/index.tsx
@@ -1,16 +1,18 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { isAxiosError } from 'axios';
 import { useNavigate } from 'react-router-dom';
 import { Job } from '../../types';
 import { deleteJob } from '../../services/jobs';
 import { useStateValue, removeJob } from '../../state';
+import useErrorHandler from '../../hooks/useErrorHandler';
 import Custom404 from '../custom404';
+import Error from '../../components/Error';
 import JobInfo from '../../components/JobInfo';
 
 const SingleJob = ({ job }: { job: Job | null | undefined }) => {
   const navigate = useNavigate();
   const [{ user }, dispatch] = useStateValue();
-  const [error, setError] = useState('');
+  const [error, handleError] = useErrorHandler();
 
   useEffect(() => {
     if (!user) {
@@ -27,19 +29,26 @@ const SingleJob = ({ job }: { job: Job | null | undefined }) => {
   }
   const handleDelete = async () => {
     try {
-      await deleteJob(user.token, job.id);
+      const userId = window.localStorage.getItem('id');
+
+      if (!userId) {
+        navigate('/');
+        return null;
+      }
+
+      await deleteJob(user.token, Number(userId), job.id);
       dispatch(removeJob(job.id));
       navigate('/jobs');
     } catch (err) {
       if (isAxiosError(err)) {
-        setError(err.response?.data);
+        handleError(err.response?.data);
       }
     }
   };
 
   return (
     <main>
-      {error && <p>{error}</p>}
+      {error && <Error err={error} />}
       <JobInfo
         job={job}
         handleDelete={handleDelete}
